--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -99,9 +99,9 @@ references:
     paths:
       - '.git'
 
-  # npm cache
+  # yarn cache
   #
-  # npm caches the modules it installs in ~/.npm.
+  # yarn caches the modules it installs in ~/.cache/yarn.
   # We cache that cache to save time pulling modules from the network.
   #
   #
@@ -110,29 +110,29 @@ references:
   #   - https://github.com/Automattic/wp-calypso/pull/25487
   #   - https://github.com/Automattic/wp-calypso/pull/27180
   #
-  # More about the CircleCI cache: https://circleci.com/docs/2.0/caching
-  restore-npm-cache: &restore-npm-cache
-    name: 'Restore npm cache'
+  # More about the CircleCI cache: https://circleci.com/docs/2.0/caching and https://circleci.com/docs/2.0/yarn/
+  restore-yarn-cache: &restore-yarn-cache
+    name: 'Restore yarn cache'
     keys:
-      - v{{ .Environment.GLOBAL_CACHE_PREFIX }}-v3-npm-modules-{{ checksum ".nvmrc" }}-{{ checksum "package-lock.json" }}
-      - v{{ .Environment.GLOBAL_CACHE_PREFIX }}-v3-npm-modules-{{ checksum ".nvmrc" }}
+      - v{{ .Environment.GLOBAL_CACHE_PREFIX }}-yarn-modules-{{ checksum ".nvmrc" }}-{{ checksum "yarn.lock" }}
+      - v{{ .Environment.GLOBAL_CACHE_PREFIX }}-yarn-modules-{{ checksum ".nvmrc" }}
 
-  npm-install: &npm-install
+  yarn-install: &yarn-install
     environment:
       PUPPETEER_SKIP_CHROMIUM_DOWNLOAD: "true"
-    name: Install npm dependencies
-    command: npm ci
+    name: Install dependencies
+    command: yarn install --frozen-lockfile
 
-  save-npm-cache: &save-npm-cache
-    name: 'Save node_modules cache'
-    key: v{{ .Environment.GLOBAL_CACHE_PREFIX }}-v3-npm-modules-{{ checksum ".nvmrc" }}-{{ checksum "package-lock.json" }}
+  save-yarn-cache: &save-yarn-cache
+    name: 'Save yarn cache'
+    key: v{{ .Environment.GLOBAL_CACHE_PREFIX }}-yarn-modules-{{ checksum ".nvmrc" }}-{{ checksum "yarn.lock" }}
     paths:
-      - ~/.npm
-  npm-e2e-install: &npm-e2e-install
-    name: Install e2e npm dependencies
+      - ~/.cache/yarn
+  yarn-e2e-install: &yarn-e2e-install
+    name: Install e2e yarn dependencies
     command: |
       cd test/e2e &&
-      CHROMEDRIVER_VERSION=$(<.chromedriver_version) npm ci
+      CHROMEDRIVER_VERSION=$(<.chromedriver_version) yarn install ----frozen-lockfile
 
   # Babel cache
   # More about the CircleCI cache: https://circleci.com/docs/2.0/caching
@@ -196,11 +196,11 @@ jobs:
       - run: *update-git-master
       - save_cache: *save-git-cache
       # npm dependencies
-      - restore_cache: *restore-npm-cache
-      - run: *npm-install
-      - run: *npm-e2e-install
-      - save_cache: *save-npm-cache
-      - run: npm run build-packages
+      - restore_cache: *restore-yarn-cache
+      - run: *yarn-install
+      - run: *yarn-e2e-install
+      - save_cache: *save-yarn-cache
+      - run: yarn run build-packages
       - persist_to_workspace:
           root: '~'
           paths:
@@ -214,7 +214,7 @@ jobs:
       - run:
           name: TypeScript typecheck
           # Report but don't fail the build (`|| exit 0`)
-          command: npm run typecheck || exit 0
+          command: yarn run typecheck || exit 0
 
   typecheck-strict:
     <<: *defaults
@@ -223,7 +223,7 @@ jobs:
       - prepare
       - run:
           name: TypeScript strict typecheck of individual subprojects
-          command: npm run tsc -- --project client/landing/gutenboarding
+          command: yarn run tsc --project client/landing/gutenboarding
 
   lint-and-translate:
     <<: *defaults
@@ -233,7 +233,7 @@ jobs:
       - run:
           name: Lint Config Keys
           when: always
-          command: npm run lint:config-defaults
+          command: yarn run lint:config-defaults
       - run:
           name: Lint Client and Server
           when: always
@@ -256,7 +256,7 @@ jobs:
           name: Build calypso-strings.pot
           when: always
           command: |
-            npm run translate
+            yarn run translate
             mv calypso-strings.pot "$CIRCLE_ARTIFACTS/translate"
       - run:
           name: Build New Strings .pot
@@ -497,7 +497,7 @@ jobs:
     steps:
       - prepare
       - run: *set-e2e-variables
-      - run: npm run decryptconfig
+      - run: yarn run decryptconfig
       - run: ./scripts/randomize.sh specs
       - run: ./scripts/run-wrapper.sh
       - run: *move-e2e-artifacts
@@ -561,7 +561,7 @@ jobs:
                   }" << parameters.slack-webhook >>
               circleci-agent step halt
             fi
-      - run: npm run decryptconfig
+      - run: yarn run decryptconfig
       - run:
           name: Run Canary Tests
           command: |