--- a/.github/workflows/full-site-editing-plugin.yml
+++ b/.github/workflows/full-site-editing-plugin.yml
@@ -4,7 +4,7 @@ on:
     paths:
     - 'apps/full-site-editing/full-site-editing-plugin/**'
 
-name: Build Full Site Editing plugin
+name: build-fse-plugin
 
 jobs:
   build:
@@ -17,7 +17,7 @@ jobs:
     - name: Checkout code
       uses: actions/checkout@master
     - name: Install dependencies
-      run: npm ci
+      run: yarn install --frozen-lockfile
     - name: Build FSE
       run: npx lerna run build --scope='@automattic/full-site-editing' --stream
     - name: Upload build artifact