--- a/.github/workflows/icfy-stats.yml
+++ b/.github/workflows/icfy-stats.yml
@@ -12,7 +12,7 @@ jobs:
       - name: Set up Node
         uses: actions/setup-node@v1
         with:
-          node-version: '>=12.15.0'
+          node-version: '^12.13.1'
       - name: Checkout code
         uses: actions/checkout@v2
       - name: Fetch git history