--- a/Dockerfile
+++ b/Dockerfile
@@ -25,7 +25,7 @@ RUN        bash /tmp/env-config.sh
 # This layer is populated with up-to-date files from
 # Calypso development.
 COPY . /calypso/
-RUN npm ci
+RUN yarn install --frozen-lockfile
 
 
 # Build the final layer
@@ -36,7 +36,7 @@ ARG        commit_sha="(unknown)"
 ENV        COMMIT_SHA $commit_sha
 
 ARG        workers
-RUN        WORKERS=$workers CALYPSO_ENV=production BUILD_TRANSLATION_CHUNKS=true npm run build
+RUN        WORKERS=$workers CALYPSO_ENV=production BUILD_TRANSLATION_CHUNKS=true yarn run build
 
 USER       nobody
 CMD        NODE_ENV=production node build/bundle.js