--- a/apps/full-site-editing/bin/sync-blog-posts-block.sh
+++ b/apps/full-site-editing/bin/sync-blog-posts-block.sh
@@ -40,6 +40,7 @@ fi
 # print usage is no mode matched
 if [ -z "$MODE" ]
 then
+<<<<<<< HEAD
 	echo "Usage: npm run sync:blog-posts-block [arguments]"
 	echo
 	echo Possible arguments:
@@ -49,6 +50,17 @@ then
 	echo You can find the latest release ID on https://github.com/Automattic/newspack-blocks/releases/latest
 	echo
 	exit 1
+=======
+    echo "Usage: yarn run sync:blog-posts-block [arguments]"
+    echo
+    echo Possible arguments:
+    echo --branch=master
+    echo --release=1.0.0-alpha.17
+    echo
+    echo You can find the latest release ID on https://github.com/Automattic/newspack-blocks/releases/latest
+    echo
+    exit 1
+>>>>>>> drop the npm lockfile
 fi
 
 TARGET=./full-site-editing-plugin/blog-posts-block/newspack-homepage-articles