--- a/apps/full-site-editing/bin/sync-blog-posts-block.sh
+++ b/apps/full-site-editing/bin/sync-blog-posts-block.sh
@@ -40,17 +40,6 @@ fi
 # print usage is no mode matched
 if [ -z "$MODE" ]
 then
-<<<<<<< HEAD
-	echo "Usage: npm run sync:blog-posts-block [arguments]"
-	echo
-	echo Possible arguments:
-	echo --branch=master
-	echo --release=1.0.0-alpha.17
-	echo
-	echo You can find the latest release ID on https://github.com/Automattic/newspack-blocks/releases/latest
-	echo
-	exit 1
-=======
     echo "Usage: yarn run sync:blog-posts-block [arguments]"
     echo
     echo Possible arguments:
@@ -60,7 +49,6 @@ then
     echo You can find the latest release ID on https://github.com/Automattic/newspack-blocks/releases/latest
     echo
     exit 1
->>>>>>> drop the npm lockfile
 fi
 
 TARGET=./full-site-editing-plugin/blog-posts-block/newspack-homepage-articles
@@ -110,7 +98,7 @@ elif [ "$MODE" = "path" ] ; then
 	CODE="${npm_config_path}"
 elif [ "$MODE" = "npm" ] ; then
 	# Way back to wp-calypso root:
-	CODE="../../node_modules/newspack-blocks" 
+	CODE="../../node_modules/newspack-blocks"
 fi
 
 if [ ! -d "$CODE" ] ; then