--- a/apps/full-site-editing/full-site-editing-plugin/blog-posts-block/README.md
+++ b/apps/full-site-editing/full-site-editing-plugin/blog-posts-block/README.md
@@ -26,7 +26,7 @@ Once your changes land on the Newspack side, coordinate with the team (over issu
 While being in `apps/full-site-editing/` directory, you can run:
 
 ```
-npm run sync:blog-posts-block --release=1.0.0-alpha.17
+yarn run sync:blog-posts-block --release=1.0.0-alpha.17
 ```
 
 This will pull the code from the release and integrate it into this repository. Please review changes, make sure to update `NEWSPACK_BLOCKS__VERSION` in [index.php](./index.php) and commit.
@@ -36,5 +36,5 @@ This will pull the code from the release and integrate it into this repository. 
 Sometimes, probably, you will need to sync the NHA code straight in your local environment. It means you will get working on both projects at the same time. For this situation, you'd like to reference the code source through the `path` bin script argument.
 
 ```js
-npm run sync:blog-posts-block --path=/Absolute/path/of/newspack-blocks/
+yarn run sync:blog-posts-block --path=/Absolute/path/of/newspack-blocks/
 ```