--- a/apps/notifications/package.json
+++ b/apps/notifications/package.json
@@ -22,7 +22,7 @@
 		"build": "calypso-build --config='./webpack.config.js'"
 	},
 	"dependencies": {
-		"@automattic/calypso-color-schemes": "^2.1.0",
-		"@automattic/calypso-polyfills": "^1.0.0"
+		"@automattic/calypso-color-schemes": "2.1.0",
+		"@automattic/calypso-polyfills": "1.0.0"
 	}
 }