--- a/apps/wpcom-block-editor/package.json
+++ b/apps/wpcom-block-editor/package.json
@@ -17,11 +17,11 @@
 	"homepage": "https://github.com/Automattic/wp-calypso",
 	"scripts": {
 		"bundle": "calypso-build --env.WP",
-		"build:dev": "npm run bundle --",
-		"build:prod": "NODE_ENV=production npm run bundle --",
+		"build:dev": "yarn run bundle --",
+		"build:prod": "NODE_ENV=production yarn run bundle --",
 		"build": "npm-run-all --parallel \"build:* -- {@}\" --",
 		"clean": "npx rimraf dist",
-		"prebuild": "npm run clean"
+		"prebuild": "yarn run clean"
 	},
 	"dependencies": {
 		"debug": "4.1.1"