--- a/client/components/community-translator/README.md
+++ b/client/components/community-translator/README.md
@@ -5,4 +5,4 @@ An asynchronously-loaded, React-based community translator.
 
 At present, it's in its most basic form, loaded behind a feature flag awaiting further development.
 
-To enable in development, run `ENABLE_FEATURES=i18n/community-translator npm start`
+To enable in development, run `ENABLE_FEATURES=i18n/community-translator yarn start`