--- a/client/state/initial-state.js
+++ b/client/state/initial-state.js
@@ -61,8 +61,8 @@ function shouldPersist() {
  * browser state and loading without it
  *
  * Can be overridden on the command-line with two flags:
- *   - ENABLE_FEATURES=force-sympathy npm start (always sympathize)
- *   - ENABLE_FEATURES=no-force-sympathy npm start (always prevent sympathy)
+ *   - ENABLE_FEATURES=force-sympathy yarn start (always sympathize)
+ *   - ENABLE_FEATURES=no-force-sympathy yarn start (always prevent sympathy)
  *
  * If both of these flags are set, then `force-sympathy` takes precedence.
  *