--- a/client/state/nps-survey/test/README.md
+++ b/client/state/nps-survey/test/README.md
@@ -2,5 +2,5 @@ NPS Survey Testing
 ==================
 
 ```
-npm run test-client -- --grep "state nps-survey"
+yarn run test-client -- --grep "state nps-survey"
 ```