--- a/client/state/push-notifications/test/README.md
+++ b/client/state/push-notifications/test/README.md
@@ -1,4 +1,4 @@
 Push Notification Testing
 ===
 
-`npm run test-client -- --grep "state push-notifications"`
+`yarn run test-client -- --grep "state push-notifications"`