--- a/client/state/ui/nps-survey-notice/test/README.md
+++ b/client/state/ui/nps-survey-notice/test/README.md
@@ -2,5 +2,5 @@ NPS Survey Notice Testing
 =========================
 
 ```
-npm run test-client -- --grep "state ui nps-survey-notice"
+yarn run test-client -- --grep "state ui nps-survey-notice"
 ```