--- a/client/state/ui/post-type-list/README.md
+++ b/client/state/ui/post-type-list/README.md
@@ -2,5 +2,5 @@ Post Type List Testing
 =========================
 
 ```
-npm run test-client client/state/ui/post-type-list/test
+yarn run test-client client/state/ui/post-type-list/test
 ```