--- a/config/README.md
+++ b/config/README.md
@@ -31,7 +31,7 @@ The config files contain a features object that can be used to determine whether
 If you want to temporarily enable/disable some feature flags for a given build, you can do so by setting the `ENABLE_FEATURES` and/or `DISABLE_FEATURES` environment variables. Set them to a comma separated list of features you want to enable/disable, respectively:
 
 ```bash
-ENABLE_FEATURES=manage/plugins/compatibility-warning DISABLE_FEATURES=reader npm start
+ENABLE_FEATURES=manage/plugins/compatibility-warning DISABLE_FEATURES=reader yarn start
 ```
 ### Testing Feature Flags via URLs