--- a/test/e2e/scripts/run-wrapper.sh
+++ b/test/e2e/scripts/run-wrapper.sh
@@ -42,4 +42,4 @@ fi
 #  TESTARGS+=" -x"
 #fi
 
-npm test
+yarn test