… .yarnrc
… apps/full-site-editing/package.json
… client/package.json
… client/server/bundler/README.md
… docs/CONTRIBUTING.md
… docs/coding-guidelines/css.md
… docs/coding-guidelines/javascript.md
… docs/dependency-management.md
… docs/development-workflow.md
… docs/guide/hello-world.md
… docs/install.md
… docs/lockfile.md
… docs/merge-checklist.md
… docs/monorepo.md
… docs/performance.md
… docs/rtl.md
… docs/testing/component-tests.md
… docs/testing/snapshot-testing.md
… docs/testing/testing-overview.md
… docs/troubleshooting.md
… lerna.json
… out
… package.json
… packages/README.md
… packages/calypso-analytics/package.json
… packages/calypso-build/README.md
… packages/calypso-codemods/README.md
… packages/calypso-color-schemes/README.md
… packages/components/README.md
… packages/components/package.json