Try out yarn!

- [x] ditch the npm package-lock
- [x] get a yarn install to actually work (needs https://github.com/Automattic/color-studio/pull/23, currently bypassing by running `yarn --ignore-engines`)
- [x] turn on workspaces
- [x] update file: refs in various package.json's to semver refs based on current monorepo package versions 
- [x] update circle build
- [x] update docker build
- [x] get e2e tests working - likely need to turn it into a workspace
- [ ] ~~update `peerDependency` warnings - this probably involve adding the necessary`devDependencies` to each of the workspaces~~ (will do this in a future PR to keep this PR readable)
- [x] check whether assumptions still hold true in dep scripts
- [ ] update renovate?
- [x] remove usage of `--ignore-engines`
- [ ] ~~decide if we want to enforce [`yarn-deduplicate`](https://github.com/atlassian/yarn-deduplicate)~~
- [x] update lerna config to use yarn and workspaces
- [ ] ...