@jameslnewell 2020-01-29
While running `npm start` I was getting errors with `wpcom` trying to use the symlinked package in `package/wpcom.js/src/index.js`. It looks like this package is basically a historical copy as there's no `file:` references and no `prepare` script to build the files. 

The `make` command in its `package.json` didn't produce workable output so I've moved the whole package from `packages/wpcom.js/` to `legacy/wpcom.js` so it `yarn` will install the published version and not try symlinking the local version.