@blowery 2020-01-29
@jsnajdr and @sgomes are working on getting wpcom.js into packages properly in some other PRs. There's also https://github.com/Automattic/wp-calypso/pull/39139 coming in that will likely follow the path same kind of path.