@jameslnewell 2020-01-30
The remaining CI failure:

```
  Module <rootDir>/node_modules/@wordpress/jest-preset-default/scripts/setup-globals.js in the setupFiles option was not found.
         <rootDir> is: /home/circleci/wp-calypso/apps/full-site-editing
```

Is due to `@wordpress/jest-preset-default` containing fixed paths instead of relying on the traditional NodeJS resolution algorithm e.g. https://github.com/WordPress/gutenberg/blob/master/packages/jest-preset-default/jest-preset.js#L8

Converting the paths to use `require.resolve` gets things working. I'll work on a PR.