@blowery 2020-01-30
`yarn-deduplicate` seems like a good idea. We could enforce it via a post-install script perhaps? Not sure how that will tie in with Renovate though...