@jameslnewell 2020-02-12
Current E2E failures look to be because I deleted the `package-lock.json` for the E2E tests but we're still relying on `npm ci` for those.