@kwight 2020-02-12
What's the thinking behind moving to Yarn? I knew there were big performance concerns, but haven't those been addressed with NPM 5+? Just curious if it's a done deal (looks like it when the PR is updating all of the docs), or part of a larger evaluation.