@blowery 2020-02-12
> What's the thinking behind moving to Yarn?

@kwight npm has been quite buggy with our monorepo setup. Just getting a working lockfile has become quite difficult and is tripping folks up. Additionally, yarn's support for workspaces makes working in the monorepo quite nice, and allows us to specify proper semver ranges for our reusable package dependencies.