@sirreal 2020-02-12
Will this force `wp-desktop` to move to yarn as well?