@jameslnewell 2020-02-12
> What's the thinking behind moving to Yarn?

Speed was another reason you mentioned @blowery. Do you have any measurements?