@noahtallen 2020-02-13
Are there any particular testing instructions for this yet? Or other things we can help out with?

Also, does this help us with verbosity when running monorepo commands? For example, to run `npm run start` in the FSE plugin, I have to do `npx lerna run dev --scope='@automattic/full-site-editing' --stream --no-prefix`

> speed

I deleted node_modules before doing each of these:
- (on master) `time npm ci`: 119.61s user 38.42s system 266% cpu **59.251 total**
- (on master) `time npm i`: 179.65s user 45.87s system 203% cpu **1:51.02 total**
- (this branch) `time yarn` yarn  131.49s user 48.44s system 277% cpu **1:04.86 total**

So about 45 seconds faster for me between the normal install commands, with `npm ci` being slightly better.

Not sure if there is an equivalent of `npm ci` in yarn :)