@jameslnewell 2020-02-13
> Are there any particular testing instructions for this yet? Or other things we can help out with?

Let me get the tests passing again 😄 

> Also, does this help us with verbosity when running monorepo commands? For example, to run npm run start in the FSE plugin, I have to do npx lerna run dev --scope='@automattic/full-site-editing' --stream --no-prefix

Not really... [`yarn workspace @automattic/full-site-editing run dev`](https://classic.yarnpkg.com/en/docs/cli/workspace/) isn't exactly much shorter.

Does `cd`ing into the directory and running `yarn run dev` work?

> Not sure if there is an equivalent of npm ci in yarn :)

`yarn install --frozen-lockfile`