@noahtallen 2020-02-13
> Does cding into the directory and running yarn run dev work?

Oh heck yeah, it does! That's gonna be so much nicer. Oh thank god we can dump the lerna command. I love how yarn can do the command by default, like `yarn test:js` or `yarn dev --sync`. I'm so happy about this. All of the same commands fail with `npm` from that directory. 

I also timed with frozen lockfile, looks like it's about the same as npm ci in that case. :)