@scinos 2020-03-02
When trying to add a new dependency to a random package (for example, `cd packages/viewport && yarn add react-select`), I get the error:

```
error An unexpected error occurred: "expected workspace package to exist for \"terser\""
```

I tried with a few different packages and dependencies and always seems to get the same error. I'm using yarn 1.21.1. Sounds like an older version could fix it (https://github.com/yarnpkg/yarn/issues/7807)

Adding it to the root with `yarn add -W` works though.