@scinos 2020-03-02
I wrote a tool (publish to npm pending) to generate the effective dependency tree by looking at what is actually installed at node_modules (so it ignores `yarn.lock` or `package-lock.json`) and it found some differences between this branch and the last common ancestor with master (b2e2caa9f1).

[diff.txt](https://github.com/Automattic/wp-calypso/files/4275923/diff.txt)

(note: the tool only look at `dependencies`, it ignores `devDependencies`, `peerDependencies`, etc).