@blowery 2020-03-06
@scinos Can you talk through the changes in https://github.com/Automattic/wp-calypso/pull/39121/commits/92a8351433ed0beeb77ae9fcb93f83375390e1f0 ? Why are we dropping all the carets?