@scinos 2020-03-09
@blowery I'm still exploring what is possible with yarn workspaces and what are the best patterns, that I'll write in a doc/RFC so we have better understanding of what is going on. I think we should use that doc to guide future decisions about packages and try to keep things consistent.

In this particular case, IMO we should avoid having different versions of a subpackage (as in, a package which is part of the monorepo) as dependencies (in other words, everybody in the repo must use the same version of any package that comes from `packages/*`). Therefore there is no point in having version ranges for those packages.

The rationale is that we guarantee that `packages/a` is always compatible with the code in `packages/b`, not with an older version of `packages/b` that is different from the code currently in master. Of course there are downsides to this approach, I'll detail them in the doc.