@scinos 2020-03-09
The doc https://github.com/Automattic/wp-calypso/pull/39963/files#diff-e3d899929328d010c7e040e1dbcb7f53